For backend it working previously for Authorization, however for frontend, it didnt because React has it owns environment. Hence, to tackle the API request from client side, we need to put headers for every request by taking the user.token context + some of the button by using handleClick will be disabled if the user is not authorized/no token

+++ client/src/components/WorkoutDetails.js
@@ -1,5 +1,6 @@
 // call the useWorkoutContext hook
 import { useWorkoutContext } from "../hooks/useWorkoutContext"
+import { useAuthContext } from "../hooks/useAuthContext"
 // call date-fns package
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
@@ -9,12 +10,23 @@ const WorkoutDetails = ({workout})=>{
 
     // grab the dispatch fron useWorkoutContext
     const { dispatch } = useWorkoutContext()
+    // call the user context data
+    const { user } = useAuthContext()
 
     // handleClick will handle the delete function backend and frontend
     const handleClick = async () => {
+
+        // if no user at all, just disable the delete button functionality
+        if(!user){
+            return
+        }
+        
         const response = await fetch('http://localhost:3002/api/workouts/' + workout._id,{
             // remember to put API method
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         
         // the json also need to wait for the response to finish execute the fetch before it can received them
